src0/requirements.txt
pillow
sys
heapq
itertools

optimasi ADA - KNAPSACK
searching AI - GREEDY BEST FIRST SECRH
